refactor: add a reusable modal component
